fixing some types in position/origin of testbed

+++ src/main/java/eu/wisebed/wiseml/model/setup/Origin.java
@@ -32,12 +32,12 @@ public class Origin implements Serializable {
     /**
      * the angle phi of system rotation.
      */
-    private Float phi;
+    private Integer phi;
 
     /**
      * the angle theta of system rotation.
      */
-    private Float theta;
+    private Integer theta;
 
     /**
      * this method returns the position x.
@@ -98,7 +98,7 @@ public class Origin implements Serializable {
      *
      * @return the phi angle.
      */
-    public Float getPhi() {
+    public Integer getPhi() {
         return phi;
     }
 
@@ -107,7 +107,7 @@ public class Origin implements Serializable {
      *
      * @param phi the phi angle.
      */
-    public void setPhi(final Float phi) {
+    public void setPhi(final Integer phi) {
         this.phi = phi;
     }
 
@@ -116,7 +116,7 @@ public class Origin implements Serializable {
      *
      * @return the theta angle.
      */
-    public Float getTheta() {
+    public Integer getTheta() {
         return theta;
     }
 
@@ -125,7 +125,7 @@ public class Origin implements Serializable {
      *
      * @param theta the theta angle.
      */
-    public void setTheta(final Float theta) {
+    public void setTheta(final Integer theta) {
         this.theta = theta;
     }
 }
