+++ src/test/java/eu/wisebed/wiseml/test/uth/ReadFileAsScenario.java
@@ -66,8 +66,8 @@ public class ReadFileAsScenario {
         origin1.setX((float)0);
         origin1.setY((float)0);
         origin1.setZ((float)0);
-        origin1.setPhi((float)0);
-        origin1.setTheta((float)0);
+        origin1.setPhi(0);
+        origin1.setTheta(0);
 
         // create new timeinfo...
         TimeInfo timeInfo = new TimeInfo();
